pyproject.toml
[tool.poetry]
name = "sdd"
version = "0.1.0"
description = "Spec Driven Development - orchestrates Claude through phased workflows"
authors = ["Matthew Tondreau"]

license = "MIT"
packages = [{include = "sdd", from = "src"}]

[tool.poetry.scripts]
sdd-util = "sdd.util_cli:cli"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1"
pyyaml = "^6.0"
python-frontmatter = "^1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
ruff = "^0.8"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["sdd"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
